+++ fuzz/fuzz_targets/deserialize.rs
@@ -8,5 +8,5 @@
 use libfuzzer_sys::fuzz_target;
 
 fuzz_target!(|data: &[u8]| {
-    let _ = macaroon::Macaroon::deserialize(data);
+    let _ = libmacaroon::Macaroon::deserialize(data);
 });